+++ CHANGES.rst
@@ -26,6 +26,27 @@ extract_1d
 
 - Exit gracefully if data is with F277W filter; avoid masking entire wavemap
   if subarray is SUBSTRIP96 [#6840]
+  
+linearity
+---------
+
+- Adding feature to process ZEROFRAME data with the linearity step. [#6782]
+
+ramp_fitting
+----------
+
+- Adding feature to use ZEROFRAME for ramps that are fully saturated, but
+  the ZEROFRAME data for that ramp is good. [#6782]
+
+refpix
+------
+
+- Adding feature to process ZEROFRAME data with the refpix step. [#6782]
+
+saturation
+----------
+
+- Adding feature to process ZEROFRAME data with the saturation step. [#6782]
 
 lib
 ---
@@ -37,12 +58,16 @@ stpipe
 
 - Log the CRDS context for pipeline and standalone step processing [#6835]
 
+superbias
+---------
+
+- Adding feature to process ZEROFRAME data with the superbias step. [#6782]
+
 tweakreg
 --------
 
 - Changed default value of ``fitgeom`` from ``'general'`` to ``'rshift'``
   at the request of CalWG. [#6838]
-
 
 1.5.0 (2022-05-05)
 ==================